// demo/src/react-admin/modules/shared/types/content-picker.ts
import { LinkTarget } from '@viaa/avo2-components';

import { ContentPickerType } from '../components/ContentPicker/ContentPicker.types';

export type PickerItemControls = 'SELECT' | 'TEXT_INPUT' | 'FILE_UPLOAD';

export interface PickerItem {
	label?: string;
	type: ContentPickerType;
	value: string;
	target?: LinkTarget;
}

export interface PickerTypeOption<T = ContentPickerType> {
	value: T;
	label: string;
	disabled?: boolean;
	picker: PickerItemControls;
	fetch?: (keyword: string | null, limit: number) => Promise<PickerItem[]>;
	placeholder?: string;
}